perf(frontend): stabilize account selector inputs, add parties benchmark
