+++ packages/frontend/src/mocks/data/stories/parties-extreme-100k/parties.ts
@@ -1,0 +1,4 @@
+import type { PartyFieldsFragment } from 'bff-types-generated';
+import { generateParties } from '../parties-extreme/parties.ts';
+
+export const parties: PartyFieldsFragment[] = generateParties(100_000, 300);
